requirements.txt
python-telegram-bot==20.3
requests
beautifulsoup4
pillow
playwright
